api/requirements.txt
tensorflow-cpu
torchvision
numpy
pandas
matplotlib
flask
seaborn
gunicorn
flask-cors
pyfcm
python-decouple